Add suppression attribute to GenerateNamespaceImports
Added a suppression attribute to bypass the analyzer warning for the GenerateNamespaceImports method in RefitInterfaceImports file. Additionally, formatted the GetImportedNamespaces method for better readability and added System.Diagnostics.CodeAnalysis namespace to the using directives. The analyzer was giving a false positive for a supposedly banned API which is not actually the case, hence the decision to suppress this specific warning.