Remove package and publish steps from build process
This commit removes the package and publish steps from the build cycle in both the build and release workflows. This change is made because packaging and publishing do not need to be part of the build process. They have been taken out to make the process cleaner and more efficient.