Add filename customization for source generated files
This diff modifies the RefitGeneratorSettings to allow for a customizable filename for the generated file. Previously, the interface name was used for the filename, which could be limiting in scenarios where different naming conventions are needed between the interface and the actual file.

Moreover, a diagnostic report's severity level has been changed to a warning, and it outputs the file content upon generation. This would help in debugging scenarios; identifying quickly what the outputted file content is. Lastly, the multipleInterfaces setting has been updated to take naming conventions into account.