Update README to reflect changes in configuration options
The README.md file in the Refitter.SourceGenerator was updated to reflect changes in the configuration options for the program. The NuGet download badge at the top of the README was removed. Fields for 'filename' and 'multipleInterfaces' were added to the configuration details to provide users with additional customization when generating Refit interfaces from OpenAPI specifications. These changes in the document help users better understand and utilize these new optional configurations.