Add SonarCloud workflow for code analysis
This commit adds a new GitHub Actions workflow file called `sonar-cloud.yml`. This workflow is designed to perform static code analysis using SonarCloud. The workflow is triggered on push to any branch or when a pull request is opened, synchronized, or reopened. It sets up the environment with JDK, Node and the necessary caching, installs SonarCloud scanner, performs code coverage conversion and finally runs the analysis. This workflow integration will aid in improving the code quality by identifying the code smells, bugs, vulnerabilities, and code duplications in the early stages.