Update coverage conversion in build workflow
The shell to run the coverage conversion was set to `pwsh` to ensure PowerShell is used. A new command was added to move all `*.coveragexml` files to the current directory. This is necessary to ensure all coverage files are found when uploading to Codecov regardless of their initial directory. The `files` property was updated in the `Upload coverage reports to Codecov` step to use the updated file location.