Simplify CodeCov workflow and improve test command
Refitted the CodeCov workflow for simplicity. The 'Build' step was removed because it was redundant as 'dotnet test' also builds the project. Also, the 'Test' command was modified to directly target the test project, which makes it more precise and potentially speeds up the process. Redundant commands listing files were also eliminated.