Add tests for [GeneratedCode] attributes
New tests have been added to the 'SwaggerPetstoreTests' in 'Refitter.Tests' to verify the functionality of the 'GenerateCode' method with respect to different versions and formats of the SwaggerPetstore specification (both JSON and YAML, V2 and V3), and different settings of the 'MultipleInterfaces' property. The aim is to ensure that the generated code includes the proper 'GeneratedCode' attribute.