Add settings file option to README
Added description of the new "--settings-file" option to the examples and options sections of the README file. This change reflects the addition of a new feature which allows users to specify a settings file, ignoring all other settings if specified. This feature increases flexibility and eases configuration under certain conditions.