Add assembly version to GenerateCommand output
A change has been made to the 'GenerateCommand.cs' file in order to enrich the console output information. The assembly version is now printed when the GenerateCommand is executed. This will aid in debugging and support scenarios, as users and developers can easily check the version of the software running.