Add OpenApiDocumentFactory in Refitter.Core
The OpenApiDocumentFactory has been included in the Refitter.Core. Its purpose is to create an OpenApiDocument asynchronously from a specified path or URL. This factory supports both JSON and YAML formats. HTTP responses are automatically decompressed if required. This should enhance the usage of RefitGeneratorSettings and streamline the creation process.