Refactor OpenApiDocument creation logic
Refactored the OpenApiDocument creation process in the CreateAsync method of the RefitGenerator class. Removed the individual IsHttp and IsYaml methods that were previously used and replaced them with a more streamlined CreateAsync method from the OpenApiDocumentFactory. This simplifies the code and enhances maintainability by delegating the document creation responsibility to a separate factory class.