Refactor RefitGenerator to use primary constructors
Removed the private constructor of RefitGenerator and moved the instantiation of CSharpClientGeneratorFactory to the Generate() method. This change allows the factory to be created only when necessary, optimizing the memory utilization and improving the performance of the application. Also, constructor of the RefitGenerator class has been made public to improve its accessibility.