Add option to skip OpenAPI spec validation
A new option '--skip-validation' was introduced in the Settings class, allowing users to bypass the OpenAPI specification validation. This change was implemented to give users more flexibility. In the GenerateCommand class, a conditional statement checking this setting was also added to alter the execution flow based on the user's preference.