Show message regarding support keys when logging is disabled
Adjusted the way the support key is handled when the logging is disabled. Previously, the support key wasn't printed at all if the logging was off. Now, a message specifying that the support key is unavailable when logging is disabled is printed instead. This serves to ensure that users are informed about the unavailability of the support key when logging is off rather than silently omitting it.