Update help documentation for --settings-file flag
The help documentation for --settings-file option in both README.md and Refitter/README.md has been updated. The specifics of this flag have been clarified to point out that it will ignore all other settings, except for the --output flag. The adjustment was needed to precisely define the behavior of this option and reduce potential confusion among users.