Check that CancellationToken is the last parameter
A new test case is introduced to check if CancellationToken is being generated last in the code. This is to ensure that our code-generation complies with the standard practices, which suggests to always put CancellationToken as the last parameter in a method. Also, RefitGeneratorSettings has been updated to include a new setting – UseCancellationTokens - to decide whether or not CancellationTokens should be included.