Update sample generated code in docs
All interfaces were changed to partial in both README.md and Refitter/README.md of the SwaggerPetstore project. This allows for easy extension of these interfaces without modifying the generated code. Such a change prepares the code for upcoming features that may require the addition of methods to these interfaces. This is particularly important for maintaining backwards compatibility as the project grows.