Add support for optional nullable parameters in GenerateCommand
This commit includes a change to the GenerateCommand where it now takes into account optional nullable parameters. This added functionality allows for more flexibility when dealing with nullable parameters. The ability to mark parameters as optional and nullable can help in scenarios where the parameter might not always be required.