Update name and declaration of interfaces in SourceGenerator Tests
The naming and declaration of the interfaces `ISwaggerPetstore` and `ISwaggerPetstoreWithOptionalParameters` in `Refitter.SourceGenerator.Tests\AdditionalFiles\Generated` have been changed. This commit includes renaming of `ISwaggerPetstore` to `ISwaggerPetstoreInterface` and adding `partial` keyword to the declaration of `ISwaggerPetstoreWithOptionalParameters` interface. These changes were necessary to provide more clarity and allow future extensions of the interface.