Add tests for skipping deprecated endpoints
Added a new test suite `DeprecatedMultipleInterfaceEndpointsSkippedTests` under `Refitter.Tests`. This class contains tests to ensure that Refitter correctly skips generation of deprecated endpoints defined in OpenAPI specifications. This behavior is important as we don't want to generate unnecessary code for deprecated operations. This change helps to keep our generated codebase clean and up to date with the latest API design.