Rearrange OpenApiPath validation checks in GenerateCommand
Adjusted the order of the if statements in the GenerateCommand class. The check for whether the OpenApiPath is a URL has been moved down, after the validation for OperationNameTemplate. This ensures that the OperationNameTemplate validation is performed first, thus providing more accurate validation error messages in situations where both the URL and the OperationNameTemplate may be invalid.