Remove serviceCollectionRegistration setting
This commit removes the serviceCollectionRegistration setting from both GenerateCommand.cs and Settings.cs. The reason for removal is to simplify the generation process by no longer generating code for registering interfaces with Microsoft.Extensions.DependencyInjection. The decision is due to the fact that such registration should be left to the consuming project, allowing for more flexibility and control.