Introduce DependencyInjectionGenerator and its tests
This commit includes the implementation of a new DependencyInjectionGenerator class which is used to generate Dependency Injection configurations for Refit clients in a service collection. Furthermore, we also added tests to verify the code produced by the generator. This change makes it easier to set up Refit clients with built-in Polly support and multiple HttpMessageHandlers.