Change parameter type and update condition in Generate method
Updated the type of the 'interfaceNames' parameter from IEnumerable<string> to string[] in the Generate method in DependencyInjectionGenerator.cs. Also, an additional condition is added in the if block to check if 'interfaceNames' is empty. This was necessary because the subsequent code execution assumes that 'interfaceNames' contains at least one item and will fail otherwise.