Make baseUrl optional
The changes were made in DependencyInjectionGenerator and associated test case. The changes allow for the base URL to be null, which makes it possible to configure Refit clients without a baseUrl. Earlier, it was mandatory to provide a baseUrl but with these changes, we can configure clients without specifying any baseUrl. These changes provide flexibility in configuring Refit clients and hence improve the usability of the library.