Update test MinimalApi project
A new class `TelemetryDelegatingHandler` is added to track outbound HTTP requests. It logs every outbound HTTP request along with its method and URL. It is registered as a transient dependency and will be added to the HTTP handlers via the `AddRefitClient` extension method. Changes have also been made in the `MinimalApi.csproj` and `Program.cs` files to reflect this update. Telemetry can be very useful to gain insights, especially about bottlenecks and to improve performance. Newline issue at the end of the files is also addressed.

[skip ci]