Add polly retry settings to petstore refitter
This commit adds two new fields to the "dependencyInjectionSettings" in the petstore refitter. 'pollyMaxRetryCount' is added with a value of 3, to specify the maximum number of retries. Also, 'firstBackoffRetryInSeconds' is added with a value of 0.5 to indicate the backoff delay before the first retry. These settings will ensure more reliable request handling by implementing retries in case of failure.

[skip ci]