Update description for operation name template option
Enhance the description for the `--operation-name-template` option in `Settings.cs`. The added explanation clarifies that when using the `--multiple-interfaces ByEndpoint` option, the operation name template will be used as the name for the `Execute()` method in the generated interface. Consequently, the same understanding is reflected in the README documentation for both the root level and Refitter library usage. This change was essential to provide users with a clearer understanding of how the `--operation-name-template` works with the `--multiple-interfaces ByEndpoint`.