Rename `ContractsGeneratorSettings` to `CodeGeneratorSettings`
The name of `ContractsGeneratorSettings` is changed to `CodeGeneratorSettings` throughout the project files. This name is more reflective of the setting's usage, as it influences the generation of all types, not just contracts. The NSwag configuration JSON property is also updated from `contractsGeneratorSettings` to `codeGeneratorSettings` to reflect this change and maintain consistency.