Add outputFolder and outputFilename to petstore.refitter
These additional fields were added to petstore.refitter to simplify the refitter configuration. It eliminates the redundancy of specifying the output file path every time in the smoke-tests script. As a result, the refitter call within the smoke-tests script has been adjusted to no longer include the output path. The output file name was also updated in the GenerateAndBuild call to match the petstore.refitter setting.