Update documentation and code comment for operationNameTemplate
Updated the READMEs and code comments in the refitter project to properly reflect the usage of the 'operationNameTemplate' setting. The previous description was unclear on its function when using multiple interfaces setting as 'ByEndpoint'. The changes provide better, clearer guidance for users on how to use the setting in different scenarios.