Update Qodana Scan arguments in GitHub workflow
Updated the arguments used in the Qodana scan in our GitHub actions workflow. Instead of simply targeting the "src" directory, we are now specifically targeting the "Refitter.sln" solution within the "src" directory. This change was made to narrow the focus of the scan, and improving the relevance and effectiveness of the scan results.