Update source generator tests to generator dependency injection settings with message handlers and polly
A Polly retry policy as well as HTTP message handlers have been added to `ConfigureRefitClients` method in `IServiceCollectionExtensions`. This configuration provides the service more robustness to temporary issues by retrying requests in case of transient HTTP errors. This was guided by the new dependencyInjectionSettings in the refit configuration file. The Polly configuration is flexible, allowing different types of backoff strategies. Here a 'DecorrelatedJitterBackoff' strategy was used with a maximum of 3 retries and initial backoff of 0.5 seconds.