Upgrade target framework to .NET 8.0
The overall target framework of the project, including Refitter, Refitter.SourceGenerator.Tests, MinimalApi and Refitter.Tests, has been updated from .NET 6.0 to .NET 8.0. This is to leverage the updated features and improvements in .NET 8.0, allowing the project to stay up-to-date with the latest technology. Following this change, the launch.json file is also modified to point to the updated .NET 8.0 framework. This will ensure the smooth operation of the project and compatibility with the .NET 8.0 version.