Add colon to PascalCase conversion
Added a new method ConvertColonsToPascalCase in the StringCasingExtensions class. This method will allow us to convert string with colons to PascalCase. The change will provide a consistent way to represent colon-separated values in PascalCase, matching our coding style convention.

Also, a new unit test method was added for it in the StringCasingExtensionTests class to ensure the functionality works as expected.