Fix source generator error - System.IO.FileNotFoundException: Could not load file or assembly 'Microsoft.Bcl.AsyncInterfaces'
This is a work around to a weird error that occurs in the source generator where deserialization fails due to a FileNotFoundException. This seems to only happen in Visual Studio, and even though the error occurs, the Refit interface is still generated.

The deserialization of settings has been refactored to use a `TryDeserialize()` method. The new method attempts to deserialize, returning null if it fails, which significantly improves error handling. Moreover, error messages have also been updated to display more specific and clear reasons for failures.