Refactor OpenApiDocumentFactory to handle HTTP URLs and YAML files
This commit refactors the `OpenApiDocumentFactory` class to handle HTTP URLs and YAML files. It adds support for reading OpenAPI documents from HTTP URLs and populating missing required fields. It also includes methods to determine if a path is an HTTP URL and to get the content of a URI as a string. This refactor improves the flexibility and robustness of the `OpenApiDocumentFactory` class.