Suggest --skip-validation on validation errors
Refactored the `GenerateCommand` class in `GenerateCommand.cs`. The suggestion to use the `--skip-validation` argument is now only displayed when `SkipValidation` is false in the settings. The line of code for this suggestion has been moved from the `if` block checking for `OpenApiUnsupportedSpecVersionException` to a new `if` block. No other changes have been made to the code.